# lib/api-zod/src/generated/types/adminAnalytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Eclipse Transport API — private luxury transportation in Los Angeles
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyRevenue } from './monthlyRevenue';
import type { ServiceBookingCount } from './serviceBookingCount';

export interface AdminAnalytics {
  totalRevenue: number;
  avgBookingValue: number;
  completedTrips: number;
  monthlyRevenue: MonthlyRevenue[];
  bookingsByService: ServiceBookingCount[];
}