# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Eclipse Transport API — private luxury transportation in Los Angeles
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  passengerName: string;
  passengerEmail: string;
  passengerPhone: string;
  pickupLocation: string;
  dropoffLocation: string;
  pickupDate: string;
  pickupTime: string;
  serviceType: string;
  passengers?: number;
  /** @nullable */
  specialRequests?: string | null;
  status: BookingStatus;
  /** @nullable */
  totalPrice?: number | null;
  createdAt: string;
}