# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Eclipse Transport API — private luxury transportation in Los Angeles
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 1 */
  passengerName: string;
  passengerEmail: string;
  /** @minLength 1 */
  passengerPhone: string;
  /** @minLength 1 */
  pickupLocation: string;
  /** @minLength 1 */
  dropoffLocation: string;
  pickupDate: string;
  pickupTime: string;
  serviceType: string;
  /**
     * @minimum 1
     * @maximum 14
     */
  passengers?: number;
  /** @minimum 0 */
  luggage?: number;
  vehicleType?: string;
  /** @minimum 0 */
  estimatedPrice?: number;
  specialRequests?: string;
}