# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Eclipse Transport API — private luxury transportation in Los Angeles
 * OpenAPI spec version: 0.1.0
 */
import type { BookingUpdateStatus } from './bookingUpdateStatus';

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  totalPrice?: number;
  specialRequests?: string;
}